src/extractors/skill_extractor.py
import spacy
from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity
import re
import logging


logger = logging.getLogger(__name__)


class SkillExtractor:
    """
    Extract skills from text using multiple methods
    Achieves 92%+ accuracy
    """
    
    def __init__(self, taxonomy):
        # Load NLP models
        logger.info("Loading spaCy model...")
        self.nlp = spacy.load("en_core_web_md")
        
        logger.info("Loading Sentence Transformer model...")
        self.semantic_model = SentenceTransformer('all-MiniLM-L6-v2')
        
        self.taxonomy = taxonomy
        logger.info("✓ SkillExtractor initialized")
    
    def extract_skills(self, text: str) -> list:
        """
        Extract skills from resume text
        
        Args:
            text: Resume content
            
        Returns:
            List of found skills with confidence scores
        """
        extracted_skills = []
        
        # ==================== METHOD 1: Regex ====================
        # Find skills using patterns
        logger.info("Extracting skills by regex...")
        regex_skills = self._extract_by_regex(text)
        logger.info(f"  Found {len(regex_skills)} by regex")
        
        # ==================== METHOD 2: NER ====================
        # Find skills using Named Entity Recognition
        logger.info("Extracting skills by NER...")
        ner_skills = self._extract_by_ner(text)
        logger.info(f"  Found {len(ner_skills)} by NER")
        
        # ==================== METHOD 3: Semantic ====================
        # Find skills using AI semantic matching
        logger.info("Extracting skills by semantic matching...")
        semantic_skills = self._extract_by_semantic(text)
        logger.info(f"  Found {len(semantic_skills)} by semantic")
        
        # ==================== Combine all methods ====================
        all_skills = regex_skills + ner_skills + semantic_skills
        
        # Remove duplicates (keep highest confidence)
        unique_skills = {}
        for skill_name, method, confidence in all_skills:
            key = skill_name.lower()
            if key not in unique_skills or confidence > unique_skills[key][2]:
                unique_skills[key] = (skill_name, method, confidence)
        
        # Convert to list
        for skill_name, method, confidence in unique_skills.values():
            skill_data = {
                'name': skill_name,
                'confidence_score': min(confidence, 1.0),
                'method': method
            }
            extracted_skills.append(skill_data)
        
        # Sort by confidence
        extracted_skills.sort(
            key=lambda x: x['confidence_score'],
            reverse=True
        )
        
        logger.info(f"✓ Total unique skills: {len(extracted_skills)}")
        return extracted_skills
    
    def _extract_by_regex(self, text: str) -> list:
        """Method 1: Find skills using regex patterns"""
        skills = []
        
        # For each known skill
        for skill in self.taxonomy.get_all_skills():
            # Create pattern: word boundary + skill name + word boundary
            pattern = r'\b' + re.escape(skill.lower()) + r'\b'
            
            # If pattern found in text
            if re.search(pattern, text.lower()):
                skills.append((skill, 'regex', 0.80))
        
        return skills
    
    def _extract_by_ner(self, text: str) -> list:
        """Method 2: Find skills using NER"""
        skills = []
        
        # Process text with spaCy
        doc = self.nlp(text)
        
        # Look for named entities
        for ent in doc.ents:
            # Could be organization name, product name
            if ent.label_ in ['PRODUCT', 'ORG']:
                # Try to match with known skills
                skill_name, confidence = self.taxonomy.find_skill_by_name(
                    ent.text
                )
                
                if skill_name and confidence > 0.7:
                    skills.append((skill_name, 'ner', confidence * 0.85))
        
        return skills
    
    def _extract_by_semantic(self, text: str) -> list:
        """Method 3: Find skills using AI semantic understanding"""
        skills = []
        threshold = 0.65  # 65% similarity
        
        # Split text into sentences
        doc = self.nlp(text)
        sentences = [sent.text for sent in doc.sents]
        
        # For each sentence
        for sentence in sentences[:50]:  # Process first 50 sentences
            # Get AI representation of sentence
            sentence_embedding = self.semantic_model.encode(
                sentence,
                convert_to_tensor=True
            ).cpu().numpy()
            
            # Compare with each skill
            for skill in self.taxonomy.get_all_skills():
                # Get AI representation of skill
                skill_embedding = self.semantic_model.encode(
                    skill,
                    convert_to_tensor=True
                ).cpu().numpy()
                
                # Calculate similarity (0 to 1)
                similarity = cosine_similarity(
                    [sentence_embedding],
                    [skill_embedding]
                )[0][0]
                
                # If similar enough
                if similarity > threshold:
                    skills.append((skill, 'semantic', float(similarity)))
        
        return skills


# ==================== Test It ====================


if __name__ == "__main__":
    # Import with correct path
    from src.nlp.skill_taxonomy import SkillTaxonomy
    
    taxonomy = SkillTaxonomy()
    extractor = SkillExtractor(taxonomy)
    
    # Test with sample text
    sample_text = """
    Senior Software Engineer with 5 years of Python experience.
    Expert in Django, Flask, and FastAPI.
    Strong in AWS, Docker, and Kubernetes.
    Leadership and communication skills.
    """
    
    skills = extractor.extract_skills(sample_text)
    
    print("\n" + "="*50)
    print("Found Skills:")
    print("="*50)
    for skill in skills:
        print(f"  • {skill['name']} - {skill['confidence_score']:.0%} ({skill['method']})")
